Fix theme not being applied in ejected iframe view
Fixes #27, which was closed prematurely.

yarn.lock had to be regenerated, because yarn was creating a broken tree of dependencies that would error during `yarn build`. I had to install the dependencies with npm, then use `yarn import` to save that known-good tree of deps as a new `yarn.lock`.