Update Themes.tsx
This fixes a minor issue that I noticed when there are many themes added (in our use case, for multiple clients / brands) The row does not wrap and the theme tabs get condensed together. Haven't bothered to raise an issue as it's such a small thing and probably quite edge case.